requirements.txt
flask-sqlalchemy==2.4.1
gunicorn==20.0.4
setuptools==41.0.0
requests
